Stop using unified memory by default
Though we have not clarified the difference between unified
memory and normal GPU memory, a couple of performance
regressions were observed:

- A significant regression was observed when GPU memory usage
  is close to the limit. It looked like some GPU memory was
  migrated to the host even though GPU should have enough
  memory.
- A tiny regression was also observed for simple element-wise
  ops even when GPU memory usage is low.

The latter can be reproduced by

https://github.com/shinh/test/blob/master/unified_mem_perf.sh
https://github.com/shinh/test/blob/master/unified_mem_perf.py

on both p100 and v100.

We might eventually re-consider enabling unified memory, but
we are using no unified memory specific feature yet and I
guess this should be an opt-in feature like cupy's.