add Kuzushiji-MNIST dataset
The Kuzushiji-MNIST dataset is designed to be an
MNIST-like dataset focusing on handwritten Japanese
characters. All data features of KMNIST are identical to
MNIST, such as number of train and test images, image
size, etc with the intent of making it easy to run any
MNIST-based analysis or benchmarks against this new
dataset.