Change the deprecated `collections`
This commit solves the Issue #6637. It removes the deprecated usage
of abc from `collections`, which will be removed in Python3.8.

The abc is now imported as `from chainer.utils import collections_abc`,
which aligns with `chainer` code base.

For more details of deprecated usage of abc in collections, please refer
to issue #6637.