Update description for resnet.prepare
The current description says that the input image must be in RGB format, whereas on line 512-513 it converts the image to RGB style. This means that if the user has already pre-converted the image, the channels would be swapped.

The current description of the method call says that the function converts it to RGB. 

Alternative fix: change the `must be RGB` to `must be BGR`.