Add tests for x and gamma/beta with different dtypes
This commit adds tests for the cases where x and gamma/beta have
different dtypes, which was not tested, but is supposed to be supported.

This commit adds float16 alone with float64 tests for x and gamma/beta,
while for the ideep case, since it only supports float32, hence all the
variables are casted to float32 in ideep tests.