Do not expose global_kernel_registry
Static variables in headers will be unique symbols, which are
shared among multiple shared objects. This means you cannot
load multiple shared objects which are linked against
ChainerX since kernel registrations in the second shared
object will fail due to the dup kernel detection.

It would be better to get rid of unique symbols created in
reduce.cuh (kMaxBlockSize), but this would be safer since this
is a constant value.