Fix onnx_chainer.replace_func.fake_as_funcnode to reconstruct return value structure
With the current implementation alt_func is allowed to return a dict,
but it is flattened into a tuple so that the consumer of the result
who expects dict does not work.

To fix the problem, the structure is explicitly saved as a skeleton
and structured value is reconstructed from the skeleton and flattened
values.

As a side effect, not only a single level of tuple/list/dict but also
arbitrary nesting of tuple/list/dict is allowed in return value.