Fix ONNX-Chainer's GetItem converter to handle -1 correctly
When -1 is given as a slice, current implementation wrongly treat it
as slice(-1, 0) which select no elements at all. We avoid the problem
by treating it as slice(-1, np.iinfo(np.int64).max) because ONNX's
Operator Schemas recommends to pass in INT_MAX for slicing to the end
of a dimension with unknown size.